requirements.txt
Django>=5.2,<6.0
django-filter>=24.2